requirements.txt
argparse==1.4.0
colorama==0.4.4
PyYAML==6.0
requests==2.26.0
